+++ apps/backend/src/modules/notifications/notification.service.ts
@@ -10,6 +10,7 @@ import { type Service, type ValueOf } from "~/libs/types/types.js";
 import {
 	type NotificationFilter,
 	NotificationStatus,
+	type NotificationType,
 } from "./libs/enums/enums.js";
 import { NotificationError } from "./libs/exceptions/exceptions.js";
 import { filterQueryParameterToNotificationType } from "./libs/maps/maps.js";
@@ -35,21 +36,14 @@ class NotificationService implements Service {
 		this.socketService = socketService;
 	}
 
-	public async checkHasUserUnreadNotifications(
-		userId: number,
-	): Promise<boolean> {
-		return await this.notificationRepository.checkHasUserUnreadNotifications(
-			userId,
-		);
-	}
-
 	public async create(
 		payload: CreateNotificationRequestDto,
 	): Promise<NotificationResponseDto> {
-		const { receiverUserId, type, userId } = payload;
+		const { actionId, receiverUserId, type, userId } = payload;
 
 		const createdNotification = await this.notificationRepository.create(
 			NotificationEntity.initializeNew({
+				actionId,
 				receiverUserId,
 				status: NotificationStatus.UNREAD,
 				type,
@@ -60,7 +54,7 @@ class NotificationService implements Service {
 		const notification = createdNotification.toObject();
 
 		this.socketService.emitMessage({
-			event: SocketEvent.NEW_NOTIFICATION,
+			event: SocketEvent.UPDATE_NOTIFICATION,
 			payload: notification,
 			receiversIds: [String(notification.userId), String(receiverUserId)],
 			targetNamespace: SocketNamespace.NOTIFICATIONS,
@@ -80,6 +74,34 @@ class NotificationService implements Service {
 		}
 
 		return await this.notificationRepository.delete(notificationId);
+	}
+
+	public async deleteByActionId(
+		actionId: number,
+		type: ValueOf<typeof NotificationType>,
+	): Promise<boolean> {
+		const notification = await this.notificationRepository.findByActionId(
+			actionId,
+			type,
+		);
+
+		if (!notification) {
+			throw new NotificationError({
+				message: ExceptionMessage.NOTIFICATION_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const notificationObject = notification.toObject();
+
+		this.socketService.emitMessage({
+			event: SocketEvent.UPDATE_NOTIFICATION,
+			payload: notification,
+			receiversIds: [String(notificationObject.receiverUserId)],
+			targetNamespace: SocketNamespace.NOTIFICATIONS,
+		});
+
+		return await this.notificationRepository.deleteByActionId(actionId, type);
 	}
 
 	public async find(notificationId: number): Promise<NotificationResponseDto> {
@@ -123,6 +145,12 @@ class NotificationService implements Service {
 		};
 	}
 
+	public async getUnreadNotificationsCount(userId: number): Promise<number> {
+		return await this.notificationRepository.getUnreadNotificationsCount(
+			userId,
+		);
+	}
+
 	public async setReadNotifications(
 		notifactionIds: number[],
 	): Promise<AllNotificationsResponseDto> {
@@ -147,11 +175,12 @@ class NotificationService implements Service {
 			});
 		}
 
-		const { receiverUserId, status, type, userId } = payload;
+		const { actionId, receiverUserId, status, type, userId } = payload;
 
 		const updatedNotification = await this.notificationRepository.update(
 			notificationId,
 			NotificationEntity.initializeNew({
+				actionId,
 				receiverUserId,
 				status,
 				type,