+++ apps/frontend/src/modules/chat-messages/slices/chat-messages.slice.ts
@@ -2,14 +2,17 @@ import { createSlice } from "@reduxjs/toolkit";
 
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
+import { type ChatGetAllItemResponseDto } from "~/modules/chats/libs/types/types.js";
 
-import { sendMessage } from "./actions.js";
+import { sendMessage, setReadChatMessages } from "./actions.js";
 
 type State = {
+	chats: ChatGetAllItemResponseDto[];
 	dataStatus: ValueOf<typeof DataStatus>;
 };
 
 const initialState: State = {
+	chats: [],
 	dataStatus: DataStatus.IDLE,
 };
 
@@ -22,6 +25,12 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.PENDING;
 		});
 		builder.addCase(sendMessage.rejected, (state) => {
+			state.dataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(setReadChatMessages.fulfilled, (state) => {
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(setReadChatMessages.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
 		});
 	},