+++ apps/frontend/src/modules/courses/slices/courses.slice.ts
@@ -8,7 +8,7 @@ import { type CourseDto } from "../libs/types/types.js";
 import { getAll, getById, getRecommended } from "./actions.js";
 
 type State = {
-	addedVendorCourseIds: string[];
+	addedVendorCourseDataStatuses: Record<string, ValueOf<typeof DataStatus>>;
 	currentCourse: CourseDto | null;
 	recommendedCourses: CourseDto[];
 	searchDataStatus: ValueOf<typeof DataStatus>;
@@ -16,7 +16,7 @@ type State = {
 };
 
 const initialState: State = {
-	addedVendorCourseIds: [],
+	addedVendorCourseDataStatuses: {},
 	currentCourse: null,
 	recommendedCourses: [],
 	searchDataStatus: DataStatus.IDLE,
@@ -56,17 +56,20 @@ const { actions, name, reducer } = createSlice({
 			state.searchDataStatus = DataStatus.REJECTED;
 		});
 		builder.addCase(userCoursesActions.add.pending, (state, action) => {
-			state.addedVendorCourseIds = [
-				...state.addedVendorCourseIds,
-				action.meta.arg.vendorCourseId,
-			];
+			const { vendorCourseId } = action.meta.arg;
+
+			state.addedVendorCourseDataStatuses[vendorCourseId] = DataStatus.PENDING;
 		});
 		builder.addCase(userCoursesActions.add.rejected, (state, action) => {
-			state.addedVendorCourseIds = state.addedVendorCourseIds.filter(
-				(vendorCourseId) => {
-					return vendorCourseId != action.meta.arg.vendorCourseId;
-				},
-			);
+			const { vendorCourseId } = action.meta.arg;
+
+			state.addedVendorCourseDataStatuses[vendorCourseId] = DataStatus.REJECTED;
+		});
+		builder.addCase(userCoursesActions.add.fulfilled, (state, action) => {
+			const { vendorCourseId } = action.meta.arg;
+
+			state.addedVendorCourseDataStatuses[vendorCourseId] =
+				DataStatus.FULFILLED;
 		});
 	},
 	initialState,
@@ -75,7 +78,10 @@ const { actions, name, reducer } = createSlice({
 		clearCourses(state) {
 			state.recommendedCourses = [];
 			state.searchedCourses = [];
-			state.addedVendorCourseIds = [];
+			state.addedVendorCourseDataStatuses = {};
+		},
+		clearCurrentCourse(state) {
+			state.currentCourse = null;
 		},
 	},
 });