+++ packages/shared/src/modules/chat-messages/libs/types/chat-message-update-request-dto.type.ts
@@ -1,4 +1,8 @@
+import { type ValueOf } from "../../../../libs/types/types.js";
+import { type MessageStatus } from "../enums/enums.js";
+
 type ChatMessageUpdateRequestDto = {
+	status: ValueOf<typeof MessageStatus>;
 	text: string;
 };